requirements.txt
# ──────────────────────────────────────────────────────────────
# PMI Risk Management Platform — Python Requirements
# Python 3.11+
# ──────────────────────────────────────────────────────────────

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.35
alembic==1.13.2

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Validation
pydantic[email]==2.9.2

# HTTP client (for AI calls)
httpx==0.27.2

# Document Processing
python-docx==1.1.2
pdfplumber==0.11.4
openpyxl==3.1.5
xlsxwriter==3.2.0

# Environment
python-dotenv==1.0.1

# Email (bundled in stdlib — no extra package needed for smtplib)
# For production email consider:
# sendgrid==6.11.0

# CORS (bundled in fastapi)

# Optional: AI-powered extraction
# openai==1.47.0   # Uncomment if using OpenAI API

# Testing (dev only)
pytest==8.3.3
httpx==0.27.2
pytest-asyncio==0.24.0

# Production server
gunicorn==23.0.0
